Refactoring app.
