+++ streamlit_mods/__init__.py
@@ -1,3 +1,3 @@
-from .session_state_helper import SessionStateHelper
+from .helpers.session_state_helper import SessionStateHelper
 from .endpoints import Endpoints
 from .app_layout import AppLayout
